requirements.txt
atlassian-python-api >= 3.41.14
